% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cameraman}
\alias{cameraman}
\title{Photograp of a cameraman}
\format{an image of class \code{cimg}}
\usage{
cameraman
}
\description{
This image is uslay used as benchmar in SR problems
}
\keyword{datasets}
