% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{char_bivariate}
\alias{char_bivariate}
\title{Performs bivariate analysis on the character variables}
\usage{
char_bivariate(df, dep_var)
}
\arguments{
\item{df}{The data frame on which bivariate analysis is to be performed}

\item{dep_var}{Specify the dependent variable of the data frame}
}
\value{
Output is a list with bivariate summary
}
\description{
For numeric dependent variable, the function outputs mean of the dependent variable for all the classes and for categorical dependent variable the function returns crosstabs
}

