% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\docType{data}
\name{data.records}
\alias{data.records}
\title{Occurrence records for a fictional species in Finland.}
\format{Matrix of longitude and latitude (two columns) of species occurrence records.}
\usage{
data(data.records)
}
\description{
Occurrence records for a fictional species in Finland.
}
\keyword{datasets}

