% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\docType{data}
\name{data.sp}
\alias{data.sp}
\title{Geographic range for a fictional species in Finland.}
\format{RasterLayer object as defined by package raster.}
\usage{
data(data.sp)
}
\description{
Geographic range for a fictional species in Finland.
}
\keyword{datasets}

