% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\docType{data}
\name{data.worldborders}
\alias{data.worldborders}
\title{World country borders.}
\format{SpatialPolygonsDataFrame.}
\usage{
data(data.worldborders)
}
\description{
World country borders.
}
\keyword{datasets}

