% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\name{eoo}
\alias{eoo}
\title{Extent of Occurrence (EOO).}
\usage{
eoo(spData)
}
\arguments{
\item{spData}{Either a matrix of longitude and latitude (two columns) of each occurrence record or a presence/absence (1/0) map as a RasterLayer object.}
}
\value{
A single value in km2.
}
\description{
Calculates the Extent of Occurrence of a species based on either records or predicted distribution.
}
\details{
EOO is calculated as the minimum convex polygon covering all known or predicted sites for the species.
}
\examples{
data(data.records)
data(data.sp)
eoo(data.records)
eoo(data.sp)
}

