% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forensic.R
\docType{data}
\name{forensic}
\alias{forensic}
\title{Forensic dataset used in AAPA RED poster 2016
(2016) Willermet, C., Daniels, J., Edgar, H., Seeing RED: A new statistical solution to an old categorical data problem. https://www.researchgate.net/publication/301777157_Seeing_RED_A_new_statistical_solution_to_an_old_categorical_data_problem}
\format{An object of class \code{data.frame} with 877 rows and 12 columns.}
\usage{
data(forensic)
}
\description{
Dental non-metric trait data collected on forensic remains from US cities. Grouped by location and ancestry.
}
\examples{
data(forensic)



dimnames(forensic)

grp <- forensic [,1]
dat <- forensic [,2:12]

}
\keyword{datasets}
