% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R, R/trace_plot.R
\name{get_rhat}
\alias{get_rhat}
\alias{plot_rhat}
\alias{get_neff}
\alias{plot_neff}
\alias{trace_plot.stanfit}
\title{Diagnostic functions}
\usage{
get_rhat(object, pars, ...)

plot_rhat(object, pars, labels...)

get_neff(object, pars, ...)

plot_neff(object, pars, labels...)

\method{trace_plot}{stanfit}(object, pars, labels, divergences = FALSE, warmup = FALSE, ...)
}
\arguments{
\item{object}{output from call to \code{\link[rstan]{sampling}}.}

\item{pars}{character vector of posterior parameter samples to be extracted.}
}
\description{
Takes as input a \code{stanfit} object from call to \code{\link[rstan]{sampling}} and extracts Rhat and neff values.
}
\examples{
require(rstan)

mdl <- "data{ int n; vector[n] x; }	parameters{ real mu; }	model{ x ~ normal(mu, 1.0);} generated quantities{ vector[n] x_sim; real x_sim_sum; for (i in 1:n) x_sim[i] = normal_rng(mu, 1.0); x_sim_sum = sum(x_sim);}\n"	
mdl <- stan_model(model_code = mdl)
n = 20
x = rnorm(n, 0, 2)

mdl.fit <- sampling(mdl, data = list(n = n, x = x), init = function() list(mu = 0), chains = 1)

get_rhat(mdl.fit, pars = "mu")
get_neff(mdl.fit, pars = "mu")

}
