% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\name{kml}
\alias{kml}
\title{Output kml files.}
\usage{
kml(spData, filename, minimum = 3)
}
\arguments{
\item{spData}{Either a matrix of longitude and latitude (two columns) of each occurrence record or a presence/absence map as a RasterLayer object.}

\item{filename}{The name of file to save, should end with .kml.}

\item{minimum}{If the number of records is lower than the minimum no polygon in drawn, only circles of about 10km radius around each point.}
}
\description{
Creates kml files for Google Maps as required by IUCN guidelines.
}

