% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_code.R
\name{model_code}
\alias{model_code}
\title{Get model code}
\usage{
model_code(id, path = ".")
}
\arguments{
\item{id}{character vector being one of slope or poly}
}
\value{
model code
}
\description{
\code{stan} model code is extracted from the \code{red} package for compilation with \code{rstan}.
}
