% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{mv_treatment_charvar}
\alias{mv_treatment_charvar}
\title{Performs missing value treatment for character variables by imputation}
\usage{
mv_treatment_charvar(df, default = T, col.del_cutoff = 0.5,
  char_var_list1 = "EMPTY_99", char_var_list2 = "EMPTY_99")
}
\arguments{
\item{df}{The data frame on which the missing value treatment needs to be performed}

\item{default}{T/F. For default = T, the function first drops the variables with missing value percentage more than col.del_cutoff and then imputes all the missing values with unknown}

\item{col.del_cutoff}{Variables with missing value percentage greater than col.del_cutoff will be dropped. Default value is 50 percent}

\item{char_var_list1}{If default = F, then user need to specify the list of variables on which the imputation needs to be performed. In char_var_list1 user can speciy the variables where NA needs to be imputed with unknown}

\item{char_var_list2}{Specify the variables for which missing values needs to imputed with mode}
}
\value{
Output is the data set with missing values imputed
}
\description{
Imputes missing values with unknown or mode. The function also deletes the columns with missing values more than the specified cutoff
}
\section{Warning}{
 Enter only categorical variables in char_var_list1 and char_var_list1
}

