% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\name{outliers}
\alias{outliers}
\title{Visual detection of outliers.}
\usage{
outliers(longlat, layers)
}
\arguments{
\item{longlat}{Matrix of longitude and latitude (two columns) of species occurrence records.}

\item{layers}{Raster* object as defined by package raster.}
}
\value{
No values are returned, two plots are drawn for visual inspection.
}
\description{
Draws plots of sites in geographical (latlong) and environmental (2-axis PCA) space.
}
\details{
Erroneous data sources or errors in transcriptions may introduce outliers that can be easily detected by looking at simple graphs of geographical or environmental space.
}

