% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-RED.R
\name{plot_RED}
\alias{plot_RED}
\title{Plot dist objcect as dendrogram}
\usage{
plot_RED(d, type = c("2D", "3D", "tree"), labels = T, ...)
}
\arguments{
\item{d}{A distance object}

\item{type}{The type of plot to make: 2D Scatterplot, 3D Scatterplot, Dendrogram}

\item{labels}{Logical value indicating whether data point should be labeled}

\item{...}{Additional arguments to pass to plot/plot3D}
}
\value{
Nothing will happen
}
\description{
Plot the result of RED() (or any dist object) in a variety of ways. For scatter plots, Multidimensional scaling (cmdscale()) is used to visualize the distance matrix, for dendrograms, Ward's clustering is used hclust(method = "ward.D2").
}
\examples{
#with Forensic dataset

data(forensic) #load sample data

str(forensic) #check the structure of the data
grades <- forensic[,2:12] #asign the grades to an object
groups <- forensic[,1] #assign the grouping factor to another

red1 <- RED(grades, groups, dis.only = TRUE)

plot_RED(red1, type = "tree")



#Create sample data
dat1 <- matrix(rnorm(100, mean = 7.5, sd = 1), nrow = 100, ncol = 10)
dat2 <- matrix(rnorm(100, mean = 2.5, sd = 2), nrow = 100, ncol = 10)
dat3 <- matrix(rnorm(100, mean = 4, sd = 1), nrow = 100, ncol = 10)

dat <- rbind(dat1, dat2, dat3)
grp <- rep(c("A","B","C"), each = 100)

#Calculate distance matrix
dis <- RED(dat, grp)

#Visualize
plot_RED(dis, type = "2D")



}
