% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\name{raster.clean}
\alias{raster.clean}
\title{Uniformize raster layers.}
\usage{
raster.clean(layers)
}
\arguments{
\item{layers}{Raster* object as defined by package raster.}
}
\value{
A Raster* object, same class as layers.
}
\description{
Crop raster layers to minimum size and uniformize NA values accross layers.
}
\details{
Excludes all marginal rows and columns with only NA values and change values to NA if they are NA in any of the layers.
}
\examples{
data(data.layers)
raster::plot(raster.clean(data.layers))
}

