% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\name{raster.long}
\alias{raster.long}
\title{Create longitude layer.}
\usage{
raster.long(layers)
}
\arguments{
\item{layers}{Raster* object as defined by package raster.}
}
\value{
A RasterLayer object.
}
\description{
Create a layer depicting longitude based on any other.
}
\details{
Using longitude (and latitude) in models may help limiting the extrapolation of the predicted area much beyond known areas.
}
\examples{
data(red.layers)
raster::plot(raster.long(red.layers))
}
