% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\name{raster.north}
\alias{raster.north}
\title{Create northness layer.}
\usage{
raster.north(layer)
}
\arguments{
\item{layer}{RasterLayer object of elevation as defined by package raster.}
}
\value{
A RasterLayer object.
}
\description{
Create a layer depicting northness based on an elevation layer.
}
\details{
Using elevation, aspect can be calculated. Yet, it is a circular variable (0 = 360) and has to be converted to northness and eastness to be useful for modelling.
}
\examples{
data(data.sp)
raster::plot(raster.north(data.sp))
}

