% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\name{records}
\alias{records}
\title{Download taxon records from GBIF.}
\usage{
records(taxon)
}
\arguments{
\item{taxon}{Taxon name.}
}
\value{
A data.frame with longitude and latitude, plus species names if taxon is above species.
}
\description{
Downloads species or higher taxon data from GBIF and outputs records with geographical coordinates.
}

