% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RED-package.R
\docType{package}
\name{RED-package}
\alias{RED}
\alias{RED-package}
\title{RED: Robust Estimator of grade Difference (RED)}
\description{
Package of functions for calculating and visualizing Robust Estimator of grade Difference (RED). RED is a nonparametric method that avoids data compression and allows missing data. Please send any comments, or questions to robustestimator@gmail.com
}
\references{
(2018) Daniels, J., McKean, J., Willermet, C., Edgar, H., Robust Estimator of Grade Differences: a new statistical solution to an old categorical data problem, Chapter 4, Cambridge University Press.

(2016) Willermet, C., Daniels, J., Edgar, H., Seeing RED: A new statistical solution to an old categorical data problem. https://www.researchgate.net/publication/301777157_Seeing_RED_A_new_statistical_solution_to_an_old_categorical_data_problem
}
\keyword{internal}
