% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\docType{data}
\name{red.distribution}
\alias{red.distribution}
\title{Geographic range for a fictional species in Finland.}
\format{RasterLayer object as defined by package raster.}
\usage{
data(red.distribution)
}
\description{
Geographic range for a fictional species in Finland.
}
\keyword{datasets}
