% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\docType{data}
\name{red.layers}
\alias{red.layers}
\title{Environmental layers for Finland.}
\format{RasterStack object as defined by package raster.}
\usage{
data(red.layers)
}
\description{
Average annual temperature, total annual precipitation, altitude and landcover for Finland (Hijmans et al. 2005, Tuanmu & Jetz 2014).
}
\references{
Hijmans, R.J., Cameron, S.E, Parra, J.L., Jones, P.G. & Jarvis A. (2005) Very high resolution interpolated climate surfaces for global land areas. International Journal of Climatology, 25: 1965-1978.

Tuanmu, M.-N. & Jetz, W. (2014) A global 1-km consensus land-cover product for biodiversity and ecosystem modeling. Global Ecology and Biogeography, 23: 1031-1045.
}
\keyword{datasets}
