% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\docType{data}
\name{red.range}
\alias{red.range}
\title{Geographic range for Hogna maderiana (Walckenaer, 1837).}
\format{RasterLayer object as defined by package raster of range for Hogna maderiana (Walckenaer, 1837), a spider species from Madeira Island.}
\usage{
data(red.range)
}
\description{
Geographic range for Hogna maderiana (Walckenaer, 1837).
}
\keyword{datasets}
