% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RED.R
\name{RED}
\alias{RED}
\title{Calculate RED distance matrix}
\usage{
RED(dat, grp)
}
\arguments{
\item{dat}{A data matrix containing observations (rows) of variables(columns)}

\item{grp}{A grouping vector to compare the observations}
}
\value{
output Returns a list containing the distance matrix, table of group sample sizes, and standardized z-scores.
}
\description{
Calculate RED (Robust Estimator of grade Difference) on a set of ordinal or continuous data. Data are first standardized by mean and sd before pairwise differences are calculated within and between groups.
}
\keyword{}
