% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\docType{data}
\name{red.records}
\alias{red.records}
\title{Occurrence records for a fictional species in Finland.}
\format{Matrix of longitude and latitude (two columns) of occurrence records for a fictional species in Finland.}
\usage{
data(red.records)
}
\description{
Occurrence records for a fictional species in Finland.
}
\keyword{datasets}
