% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\name{red.setDir}
\alias{red.setDir}
\title{Setup GIS directory.}
\usage{
red.setDir()
}
\description{
Setup directory where GIS files are stored.
}
\details{
Writes a txt file in the red directory allowing the package to always access the world GIS files directory.
}
