% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RED_dist.R
\name{RED_dist}
\alias{RED_dist}
\title{Calculate RED distance matrix}
\usage{
RED_dist(dat, grp, dis.only = TRUE)
}
\arguments{
\item{dat}{A data matrix containing observations (rows) of variables(columns)}

\item{grp}{A grouping vector to compare the observations}

\item{dis.only}{Logical value indicating whether only the distance matrix is or a list containing the distance matrix, table of group sample sizes, and standardized z-scores.}
}
\value{
Depending on dis.only, RED() returns either a distance matrix only (dis.only = TRUE) or a list containing group sample size and standardized z-scores in addition to the distance matrix
}
\description{
Calculate RED (Robust Estimator of grade Difference) on a set of ordinal or continuous data. Data are first standardized by mean and sd before pairwise differences are calculated within and between groups.
}
\examples{


#with Forensic dataset

dat1 <- forensic #load sample data


grades1 <- dat1[,2:12] #asign the grades to an object
groups1 <- dat1[,1] #assign the grouping factor to another

dis1 <- RED_dist(grades1, groups1, dis.only = TRUE)

#Visualize
RED_plot(dis1, type = "tree")


RED_plot(dis1, type = "3D", col = rainbow(length(labels(dis1))), size = 10)

#Create sample data
dat.a <- matrix(rnorm(100, mean = 7.5, sd = 1), nrow = 100, ncol = 10)
dat.b <- matrix(rnorm(100, mean = 2.5, sd = 2), nrow = 100, ncol = 10)
dat.c <- matrix(rnorm(100, mean = 4, sd = 1), nrow = 100, ncol = 10)

dat2 <- rbind(dat.a, dat.b, dat.c)
grp2 <- rep(c("A","B","C"), each = 100)

#Calculate distance matrix
dis2 <- RED_dist(dat2, grp2)

#Visualize
RED_plot(dis2, type = "2D", pch = 15:17, col = 3:5)


}
