% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RED_pairdiff.R
\name{RED_pairdiff}
\alias{RED_pairdiff}
\title{Calculate pairwise difference between two vectors}
\usage{
RED_pairdiff(v1, v2)
}
\arguments{
\item{v1}{A numeric vector}

\item{v2}{A numeric vector}
}
\value{
Returns symmetrical pairwise difference between vectors
}
\description{
Calculate pairwise differenec between any two vectors. Calculating on the same vector twice will produce a symmetrical matrix with 0s on the diagonal
}
\examples{
a <- 1:10
b <- 5:15

RED_pairdiff(a,b)
}
