% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-helper.R
\name{RED_zscore}
\alias{RED_zscore}
\title{Standardize data by calculating z-scores.}
\usage{
RED_zscore(dat)
}
\arguments{
\item{dat}{A numeric vector to be scaled}
}
\value{
output Returns a vector of z-scores
}
\description{
Rescale a vector based on the mean, sd of the data.
}
