% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{replace_charvars}
\alias{replace_charvars}
\title{Performs variable transformations on character variables}
\usage{
replace_charvars(df, col, option = "other", cutoff = 0.05)
}
\arguments{
\item{df}{The data frame on which variable transformation is to be performed}

\item{col}{Variable names on which the transformation is to be performed}

\item{option}{Specify other to group classes with frequency less than the cutoff as other or specify rep_max to replace minority classes with modal class}

\item{cutoff}{Classes with frequency less than the cutoff will be modified. Default is 5 percent}
}
\value{
Output is the data set with transformed variables
}
\description{
Groups minority classes as others or replaces them with the modal class
}

