% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{saycolour}
\alias{saycolour}
\title{Say Colour}
\usage{
saycolour()
}
\description{
This function says the colour.
What colour is says depends on the package.
If it's the blue package then it says 'blue'.
If it's the red package then it says 'red'.
}
