% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{shift}
\alias{shift}
\title{shifting operator}
\usage{
shift(im, s)
}
\arguments{
\item{im}{cimg object}

\item{s}{numeric \code{p} by \code{2} matrix containing the registration
sameters}
}
\value{
shifted cimg object
}
\description{
shifting operator
}
