% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RED_pkg.R
\docType{package}
\name{SRI}
\alias{SRI}
\alias{SRI-package}
\title{SRI: Super-resolution imaging package}
\description{
Super-resolution imaging (SR) is a class of techniques that enhance the resolution of an imaging system.
}
