% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.map}
\alias{summary.map}
\title{Summary function for maximum a posteriori estimate}
\usage{
\method{summary}{map}(object, pars, ...)
}
\arguments{
\item{object}{return from a call to \code{\link{map}}}

\item{pars}{parameters to be summarised}
}
\description{
Takes as argument a return from a call to \code{\link{map}} and summarises the mean, sd and quantile values.
}
\note{
This function will fail of the \code{draws} argument is not specified in the optimisation i.e. \code{optimizing(..., draws = <n>)}, where \code{<n>} is a large number.
}
\examples{
require(rstan)

mdl <- "data{ int n; vector[n] x; } parameters{ real mu; real sigma;} model{ x ~ normal(mu, sigma);} generated quantities{ vector[n] x_sim; for (i in 1:n) x_sim[i] = normal_rng(mu, sigma);} \n"	
mdl <- stan_model(model_code = mdl)
n <- 20
x <- rnorm(n, 0, 2)

mdl.fit <- optimizing(mdl, data = list(n = n, x = x), init = list(mu = 0, sigma = 1), draws = 2000)

mdl.map <- map(mdl.fit, pars = c("mu", "sigma", "x_sim"), dims = list("mu" = 0, "sigma" = 0, "x_sim" = n))

summary(mdl.map, pars = c("mu", "sigma"))

cbind(x = x, x_sim = summary(mdl.map, pars = "x_sim")[[1]][, "mean"])

}
