% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Plot MCMC sample traces}
\usage{
trace_plot(object, pars, ...)
}
\arguments{
\item{object}{output from call to \code{\link[rstan]{sampling}}.}

\item{pars}{character vector of posterior parameter samples to be extracted.}

\item{labels}{character or expression vector of labels}

\item{divergences}{should divergent samples be marked?}

\item{warmup}{should warmup samples be included?}
}
\description{
Takes as input a \code{stanfit} object from call to \code{\link[rstan]{sampling}} and plots traces for desired parameters.
}
\examples{
require(rstan)

mdl <- "data{ int n; vector[n] x; }	parameters{ real mu; }	model{ x ~ normal(mu, 1.0);} generated quantities{ vector[n] x_sim; real x_sim_sum; for (i in 1:n) x_sim[i] = normal_rng(mu, 1.0); x_sim_sum = sum(x_sim);}\n"	
mdl <- stan_model(model_code = mdl)
n = 20
x = rnorm(n, 0, 2)

mdl.fit <- sampling(mdl, data = list(n = n, x = x), init = function() list(mu = 0), chains = 1)

trace_plot(mdl.fit, pars = "mu", labels = expression(mu))

}
