% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{univariate_numvar}
\alias{univariate_numvar}
\title{Performs univariate analysis on all the numerical variables of the data}
\usage{
univariate_numvar(df, outlier_cutoff = 3)
}
\arguments{
\item{df}{The data frame on which univariate is to be performed}

\item{outlier_cutoff}{This is used to calculate the boundary beyond which an observation will be flagged as outlier. Default value is 3. User can change it to 1.5, 2, 2.5 etc. Formula to calculate outlier boundry: Upper outlier boundary =  (upper quartile)+(outlier_cutoff*IQR) , Lower outlier boundary =  (lower quartile)-(outlier_cutoff*IQR).}
}
\value{
The output is a table which includes various descriptive stats for the variables
}
\description{
The function automatically identifies all the numeric variables in the data set and performs univariate analysis
}
\section{Warning}{
 Make sure all the variables in the data set have been assigned to the proper class
}

