% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{waic}
\alias{waic}
\alias{waic.matrix}
\alias{waic.list}
\title{Calculate ELPD and WAIC}
\usage{
waic(object, ...)

\method{waic}{matrix}(object, ...)

\method{waic}{list}(object, ...)
}
\arguments{
\item{object}{log-likelihood matrix or list of matrices}

\item{...}{optional log-likelihood matrices for competing models}
}
\description{
Preferred model has the highest expected log-predictive density (ELPD) value and the lowest WAIC
}
