% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\docType{data}
\name{worldborders}
\alias{worldborders}
\title{World country borders.}
\format{
SpatialPolygonsDataFrame.
}
\usage{
data(worldborders)
}
\description{
World country borders.
}
\keyword{datasets}
