% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-score.R
\name{z.score}
\alias{z.score}
\title{Standardize data by calculating z-scores.}
\usage{
z.score(dat)
}
\arguments{
\item{dat}{A numeric vector to be scaled}
}
\value{
output Returns a vector of z-scores
}
\description{
Rescale a vector based on the mean, sd of the data.
}
\keyword{}
