% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\name{eoo}
\alias{eoo}
\title{Extent of Occurrence (EOO).}
\usage{
eoo(spData)
}
\arguments{
\item{spData}{One of three options: 1) matrix of longitude and latitude (two columns) of each occurrence record; 2) matrix of easting and northing (two columns, e.g. UTM) of each occurrence record in meters;  3) RasterLayer object of predicted distribution (0/1 values).}
}
\value{
A single value in km2.
}
\description{
Calculates the Extent of Occurrence of a species based on either records or predicted distribution.
}
\details{
EOO is calculated as the minimum convex polygon covering all known or predicted sites for the species.
}
\examples{
data(red.records)
data(red.range)
eoo(red.records)
eoo(red.range)
}
