% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-helper.R
\docType{data}
\name{forensic}
\alias{forensic}
\title{Forensic sample of dental non-metric trait scores}
\usage{
data(forensic)
}
\description{
Forensic sample of dental non-metric trait scores
}
\examples{
data(forensic)


dimnames(forensic)

grp <- forensic [,1]
dat <- forensic [,2:12]

forensic
}
\keyword{datasets}
