% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-helper.R
\docType{data}
\name{forensic}
\alias{forensic}
\title{Forensic sample of dental non-metric trait scores}
\format{An object of class \code{data.frame} with 877 rows and 12 columns.}
\usage{
forensic
}
\description{
Forensic sample of dental non-metric trait scores
}
\keyword{datasets}
