% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-diff.R
\name{pair.diff}
\alias{pair.diff}
\title{Calculate pairwise difference between two vectors}
\usage{
pair.diff(v1, v2)
}
\arguments{
\item{v1}{A numeric vector}

\item{v2}{A numeric vector}
}
\value{
Returns symmetrical pairwise difference between vectors
}
\description{
Calculate pairwise differenec between any two vectors. Calculating on the same vector twice will produce a symmetrical matrix with 0s on the diagonal
}
\examples{
a <- 1:10
b <- 5:15

pair.diff(a,b)
}
\keyword{}
