% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-RED.R
\name{plot_RED}
\alias{plot_RED}
\title{Plot dist objcect as dendrogram}
\usage{
plot_RED(d, type = c("2D", "3D", "tree"), labels = T)
}
\arguments{
\item{d}{A distance object}

\item{type}{The type of plot to make: 2D Scatterplot, 3D Scatterplot, Dendrogram}

\item{labels}{Logical value indicating whether data point should be labeled}
}
\value{
Nothing will happen
}
\description{
Plot the result of RED() (or any dist object) as a dendrogram. Increased functionality to come
}
