% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{raster.lat}
\alias{raster.lat}
\title{Create latitude layer.}
\usage{
raster.lat(layers)
}
\arguments{
\item{layers}{Raster* object as defined by package raster.}
}
\value{
A RasterLayer object.
}
\description{
Create a layer depicting latitude based on any other.
}
\details{
Using latitude (and longitude) in models may help limiting the extrapolation of the predicted area much beyond known areas.
}
\examples{
layers = red.examples("red.layers")
terra::plot(raster.lat(layers[[1]]))
}
