% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{red.getDir}
\alias{red.getDir}
\title{Read GIS directory.}
\usage{
red.getDir()
}
\description{
Read directory where GIS files are stored.
}
\details{
Reads a txt file pointing to where the world GIS files are stored.
}
