% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\name{RED}
\alias{RED}
\title{REgularization by Denoising}
\usage{
RED(y, lambda, sigma, mu = NULL, functional = "SR", engine = "MF",
  niter = 50, args = NULL)
}
\arguments{
\item{y}{cimg object with the low resolution frame(s)}

\item{lambda, sigma}{numeric indicating the regularization parameters}

\item{mu}{numeric indicating the step size}

\item{functional}{character with the optimization task or function with the functional to be used}

\item{engine}{character indicating the denoised engine or function with the denoiser engine to be used}

\item{niter}{numeric indicating the maximum number of iterations}

\item{args}{arguments to be passed implicity to H HT and f}
}
\description{
REgularization by Denoising
}
\examples{

y <- degrade(lenna, noise = 0.05)
x <- RED(y, sigma = 1, lambda = 5, mu = 0.1, 'DN', niter = 10)
par(mfrow = c(1,2), mar = c(0,0,2,0)+0.1)
plot(y, interp = FALSE, axes = FALSE, main = 'Degraded Lenna')
mtext(MSE(y, lenna), side = 1, line = -2)
plot(x, interp = FALSE, axes = FALSE, main = 'Restored Lenna')
mtext(MSE(x, lenna), side = 1, line = -2)

}
