% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\docType{data}
\name{red.records}
\alias{red.records}
\title{Occurrence records for Hogna maderiana (Walckenaer, 1837).}
\format{Matrix of longitude and latitude (two columns) of occurrence records for Hogna maderiana (Walckenaer, 1837), a spider species from Madeira Island.}
\usage{
data(red.records)
}
\description{
Occurrence records for Hogna maderiana (Walckenaer, 1837).
}
\keyword{datasets}
