% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red.R
\name{red.setup}
\alias{red.setup}
\title{Download and setup GIS files.}
\usage{
red.setup()
}
\description{
Setup red to work with species distribution modelling and layers available online.
}
\details{
Please check that you have at least 50Gb free in your disk (and a fast internet connection) to download all files. In the end of the process "only" 17.4Gb will be left though. This function will:
1. Check if maxent.jar is available in the dismo package directory.
2. Ask user input for GIS directory.
3. Download global bioclim and elevation files (20) from http://biogeo.ucdavis.edu/data/worldclim/v2.0/tif/base/wc2.0_30s_bio.zip.
4. Download landcover files (12) from http://data.earthenv.org/consensus_landcover/without_DISCover/.
5. Unzip all files and delete the originals.
6. Create a new layer (1) with the dominant land cover at each cell.
7. Resample all files (33) to approximately 10x10km (for use with widespread species) grid cells.
Sit back and enjoy, this should take a while.
}
